PlayerState was moved to domain/models; state inside players was replaced from livedata to stateflow
