+++ app/src/main/java/bes/max/playlistmaker/domain/models/PlayerState.kt
@@ -1,0 +1,8 @@
+package bes.max.playlistmaker.domain.models
+
+sealed interface PlayerState {
+    object STATE_DEFAULT : PlayerState
+    object STATE_PREPARED : PlayerState
+    object STATE_PLAYING : PlayerState
+    object STATE_PAUSED : PlayerState
+}
